Update qui marche (incomplet)

+++ exodecorateur_angryballs/maladroit/modele/ObjetAvecEffetBloque.java
@@ -9,12 +9,11 @@ public class ObjetAvecEffetBloque extends ObjetAvecEffetCollision {
 	}
 
 	@Override
-	public void collisionContour(double abscisseCoinHautGauche, double ordonnéeCoinHautGauche, double largeur,
+	public void collisionContour(int abscisseCoinHautGauche, int ordonnéeCoinHautGauche, double largeur,
 			double hauteur) {
 		Collisions.collisionBilleContourAvecArretHorizontal(this.getPosition(), this.getRayon(), this.getVitesse(),
 				abscisseCoinHautGauche, largeur);
 		Collisions.collisionBilleContourAvecArretVertical(this.getPosition(), this.getRayon(), this.getVitesse(),
-				ordonnéeCoinHautGauche, hauteur);
+				ordonnéeCoinHautGauche, hauteur); 
 	}
-
 }
