+++ exodecorateur_angryballs/maladroit/modele/ObjetAvecEffetRebond.java
@@ -9,7 +9,7 @@ public class ObjetAvecEffetRebond extends ObjetAvecEffetCollision {
 	}
 
 	@Override
-	public void collisionContour(double abscisseCoinHautGauche, double ordonnéeCoinHautGauche, double largeur,
+	public void collisionContour(int abscisseCoinHautGauche, int ordonnéeCoinHautGauche, double largeur,
 			double hauteur) {
 		Collisions.collisionObjetContourAvecRebond(this.getPosition(), this.getRayon(), this.getVitesse(),
 				abscisseCoinHautGauche, ordonnéeCoinHautGauche, largeur, hauteur);
